requirements.txt
pandas>=2.0.0
great-expectations==0.18.19
pyyaml>=6.0
python-dotenv>=1.0.0
google-genai>=0.1.0
openai>=1.0.0
streamlit>=1.28.0
plotly>=5.18.0